dev-requirements.txt
pytest
pytest-nice-parametrize
